cleanup ucan module
